PaddleIP_LAP/requirements.txt
typing_extensions==4.3.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
tqdm==4.64.1
mediapipe==0.8.11
face-alignment==1.3.5
librosa==0.9.2
scipy==1.2.3
lws==1.2.7
tensorboardX==2.5.1
piq==0.7.0
paddlepaddle-gpu==2.5.2
